Moved contact.html, sample.html and quotation.html to newly created contact app and setup views and urls there
