+++ home/urls.py
@@ -19,7 +19,4 @@ urlpatterns = [
     path('team/', views.team, name='team'),
     path('eigen_drukkerij/', views.printing_house, name='printing_house'),
     path('ons_concept/', views.our_concept, name='our_concept'),
-    path('contact/', views.contact, name='contact'),
-    path('monstermap/', views.sample_kit, name='sample_kit'),
-    path('offerte/', views.quotation, name='quotation'),
 ]
