+++ postfly_jouw_online_drukkerij/settings.py
@@ -52,6 +52,7 @@ INSTALLED_APPS = [
     'cart',
     'checkout',
     'profiles',
+    'contact',
 
     # Other
     'crispy_forms',
